fetch channel on route change + hour format to 24hs + update max_size + rename logo asset in lowercase
