+++ helpers/constants.js
@@ -1,0 +1,1 @@
+export const MAX_LENGTH = 105
